requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
alembic==1.12.1
python-dotenv==1.0.0
bcrypt==4.0.1
email-validator==2.1.0.post1
supabase==1.0.3
pytest==7.4.3
httpx==0.25.1
pytest-asyncio==0.21.1
pytest-cov==4.1.0